Address cloze formatter review comments
